requirements.txt
speedtest-cli
pytest
coverage
pyinstaller
